+++ src/main/java/nextstep/ladder/factory/LadderDataConverter.java
@@ -1,0 +1,20 @@
+package nextstep.ladder.factory;
+
+import java.util.List;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
+import nextstep.ladder.generator.GeneratorWidthSize;
+import nextstep.ladder.generator.NonConsecutiveFlagGenerator;
+import nextstep.ladder.interfaces.Lines;
+
+class LadderDataConverter {
+    static List<Lines> convert(int height, int width, NonConsecutiveFlagGenerator generator) {
+        return toLadder(height, width, generator);
+    }
+
+    private static List<Lines> toLadder(int height, int width, NonConsecutiveFlagGenerator generator) {
+        return Stream.generate(() -> LinesFactory.lines(generator.create(new GeneratorWidthSize(width - 1))))
+                .limit(height)
+                .collect(Collectors.toList());
+    }
+}
