+++ src/main/java/nextstep/ladder/factory/LinesFactory.java
@@ -2,14 +2,14 @@ package nextstep.ladder.factory;
 
 import java.util.Arrays;
 import java.util.List;
-import nextstep.ladder.domain.Lines;
+import nextstep.ladder.domain.LinesImpl;
 
 public class LinesFactory {
-    public static Lines lines(List<Boolean> lines) {
-        return new Lines(LinesDataConverter.convert(lines));
+    public static LinesImpl lines(List<Boolean> lines) {
+        return new LinesImpl(LinesDataConverter.convert(lines));
     }
 
-    public static Lines lines(Boolean... lines) {
+    public static LinesImpl lines(Boolean... lines) {
         return lines(toList(lines));
     }
