+++ src/main/java/nextstep/ladder/interfaces/Lines.java
@@ -1,0 +1,9 @@
+package nextstep.ladder.interfaces;
+
+import java.util.List;
+
+public interface Lines {
+    void move(Position position);
+
+    List<Boolean> getLines();
+}
